+++ index.html
@@ -2,11 +2,16 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <link rel="icon" type="image/svg+xml" href="/tabIcon.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>My portfolio</title>
+    <title>Portfolio | Ernestas Drungilas</title>
+    <meta
+      name="description"
+      content="Portfolio page containing information about Ernestas Drungilas"
+    />
   </head>
   <body>
+    <noscript> You need to enable JavaScript to run this app. </noscript>
     <div id="root"></div>
     <script type="module" src="/src/main.tsx"></script>
   </body>
