+++ src/data/projects.ts
@@ -10,24 +10,28 @@ export interface IProject {
 const projectList: IProject[] = [
   {
     image: "page.jpg",
-    title: "Notes",
-    text: "text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text",
+    title: "Notes App",
+    text: "Main focus of this app is to let users create notes. It is possible to create private notes, however, there is ability for group of people to create and track same notes (users can join groups where they can see content which was made there). Created notes can be sorted by creation/edited date or a set tag. Users are authentication using JWT. There is also ability to change style to dark/light mode.",
     technologies: [
       "TypeScript",
       "React",
       "Styled-Components",
-      "Node",
+      "React-Router",
+      "Redux",
+      "Vite",
+      "ESLint",
+      "Node.js",
       "Express",
       "MongoDB",
     ],
     liveLink: "",
-    gitLink: "",
+    gitLink: "https://github.com/ernestasdrun/notes-proj",
   },
   {
     image: "page.jpg",
     title: "Notes",
     text: "text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text",
-    technologies: ["Vue", "JavaScript"],
+    technologies: ["JavaScript"],
     liveLink: "",
     gitLink: "",
   },
@@ -35,7 +39,7 @@ const projectList: IProject[] = [
     image: "page.jpg",
     title: "Notes",
     text: "text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text",
-    technologies: ["Vue", "JavaScript"],
+    technologies: ["JavaScript"],
     liveLink: "",
     gitLink: "",
   },
@@ -43,7 +47,7 @@ const projectList: IProject[] = [
     image: "page.jpg",
     title: "Notes",
     text: "text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text text",
-    technologies: ["Vue", "JavaScript"],
+    technologies: ["JavaScript"],
     liveLink: "",
     gitLink: "",
   },
